requirements.txt
Bottleneck @ file:///work/ci_py311/bottleneck_1676827011612/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
cryptography @ file:///croot/cryptography_1694444244250/work
greenlet @ file:///work/ci_py311/greenlet_1676825247229/work
idna @ file:///work/ci_py311/idna_1676822698822/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp311-cp311-linux_x86_64.whl#sha256=184178a893ff0b5e0cb1b0b7a0dddff49dc91424c4e4a01e29bcb05f27a698a6
pandas @ file:///croot/pandas_1697477158611/work/dist/pandas-2.1.1-cp311-cp311-linux_x86_64.whl#sha256=ed5518777956b76c6290791e98ec9ff6ea64a621c757052a7697aa4252bc0d9b
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pymongo==3.12.0
PyMySQL @ file:///work/ci_py311/pymysql_1676845339962/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv @ file:///work/ci_py311/python-dotenv_1676842320650/work
pytz @ file:///croot/pytz_1695131579487/work
requests @ file:///croot/requests_1690400202158/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///croot/sqlalchemy_1695720894144/work
tqdm @ file:///croot/tqdm_1679561862951/work
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1698257533958/work
